:zap:Feat: modify currency exchange rate func 전역변수, 커스텀 훅으로 환율 스위치 성능 개선 위젯 셀렉터에서 가격 표시 안 되던 문제 해결 Resolves: #19
