+++ coin/src/components/units/main/Main.presenter.tsx
@@ -1,6 +1,6 @@
 import * as S from "./Main.styles";
 import { useRecoilState } from "recoil";
-import { darkMode } from "../../commons/atoms";
+import { darkMode, currencyKRW } from "../../commons/atoms";
 import { useSidebar } from "../../commons/sidebar/SidebarContext";
 import { FaPlus } from "react-icons/fa";
 import Widget from "./widget/Widget";
@@ -14,6 +14,7 @@ import { faDollar } from "@fortawesome/free-solid-svg-icons";
 
 export default function MainPageUI(props: IMainPageUIProps): JSX.Element {
     const [isDarkMode] = useRecoilState(darkMode);
+    const [isCurrencyKRW, setIsCurrencyKRW] = useRecoilState(currencyKRW);
     const { sidebarOpen } = useSidebar();
 
     return (
@@ -21,8 +22,8 @@ export default function MainPageUI(props: IMainPageUIProps): JSX.Element {
             <S.Container $darkMode={isDarkMode}>
                 <S.CurrencyToggleContainer sidebarOpen={sidebarOpen}>
                     <Switch
-                            onChange={() => props.setIsCurrencyKRW(prev => !prev)}
-                            checked={props.isCurrencyKRW}
+                            onChange={() => setIsCurrencyKRW(prev => !prev)}
+                            checked={isCurrencyKRW}
                             offColor="#add836"
                             onColor="#add836"
                             uncheckedIcon={<FontAwesomeIcon icon={faWon} style={{ color: 'green', padding: '3.5px', paddingTop:'6px', fontSize: '14px'}} />}
@@ -45,7 +46,7 @@ export default function MainPageUI(props: IMainPageUIProps): JSX.Element {
                             setMenuOpen={props.setMenuOpen}
                             moveWidget={props.moveWidget}
                             onClickWidget={(symbol) => props.setSelectedSymbol(symbol)}
-                            isCurrencyKRW={props.isCurrencyKRW}
+                            isCurrencyKRW={isCurrencyKRW}
                             exchangeRate={props.exchangeRate}
                         />
                     ))}
